% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anmole.R
\name{sgraph_histogram}
\alias{sgraph_histogram}
\title{exports a histogram}
\usage{
sgraph_histogram(x, x.title, title, fillcolor, colcolor)

sgraph_histogram(x, x.title, title, fillcolor, colcolor)
}
\arguments{
\item{symbol}{}
}
\description{
this function exports a histogram

this function exports a histogram with color options
}
\examples{
 sgraph_histogram(iris$Sepal.Length,"Sepal Length","Histogram for sepal length")

 sgraph_histogram(iris$Sepal.Length,"Sepal Length","Histogram for sepal length","red","blue")

}
